Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(1 inline comment)

I'm having a hard time convincing myself that this is useful.  The only time I can think that this would be useful is that a user has accidentally initialized a repo client in their home directory.  In that case, init runs quick enough and they can fix the issue by deleting the .repo directory.

I guess the other option is that the user *wants* to nest repo clients?  If that's what you are aiming for, in what cases would a user want to nest repo clients?

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0