Update patch set 2
Patch Set 2:

If this is only for people that accidentally initialized a repo client in their homedir, let's just print a message at the end of the init.

In other words, if they initialized repo in a directory that isn't the current one, print

repo has been initialized in <other/dir/path>.
  If this is not what you wanted, please run:
  rm -r <other/dir/path/.repo>
  and try again.

I'd prefer to avoid any unnecessary prompting if possible.

Patch-set: 2
Label: Code-Review=0
Label: Verified=0