Update patch set 3
Patch Set 3:

Sorry for the bad rebase.

> There's already handling for git-remote-persistent-https not existing. How about adding handling for when exists, but does not recognise the extra flag?

The current behavior of _ConnectProxy when git-remote-persistent-https fails is to print the contents of stderr and then not use the proxy. This means if there is an unrecognized flag (i.e. -print_config to an old version), it will print usage and then proceed normally. I considered adding logic to not print the contents of stderr only if it looks like a usage error, but that seemed a bit fragile.

Another alternative is to check the build label of git-remote-persistent-https like we do for git in various places. The problem with this is that git-remote-persistent-https doesn't have a version number per se, so this isn't stable across machines. The build label has no other guarantees than two binaries built at different times have different labels. (It is only used by the proxy to determine whether the background process needs to be restarted.)

Patch-set: 3
Label: Code-Review=0