Update patch set 4
Patch Set 4: I would prefer that you didn't submit this; Verified

(7 inline comments)

This is a nice feature that I think will be really useful.

I've tested it and it works as expected, but there are a few nits/issues in the implementation and documentation.

Patch-set: 4
Label: Code-Review=-1
Label: Verified=+1