Show proper path of local manifests in deprecation warning
When a local_manifest.xml file is present, a deprecation warning
is printed telling the user to put local manifest files in the
local_manifests directory.

Include the name of the repo directory (".repo/") as a prefix on
the directory name mentioned in the warning, to reduce confusion
about where the directory should be located.

Change-Id: I85710cfbd6e77fb2fa6b7b0ce66d77693ccd649f