Allow user to specify local manifests directory on init
In the current implementation, the user is limited to including
local manifests in the local_manifests directory within the .repo
directory.

The user has to explicitly copy the local manifest files into this
location.  If the client workspace is removed, the local manifest
files are also removed.  If the same local manifests should be used
in several client workspaces, they have to be explicitly copied into
each workspace.

Add a new option on the init command, --local-manifests-dir, that
allows the user to specify the directory in which local manifests
are stored.

Change-Id: Id3ed119301bf7d207adb64247645cab019b7828b