Update patch set 2
Patch Set 2:

Sorry for the late response, I haven't had much time to work on this.

I've just tested and it works for me.

There are a couple of things that you can check to make sure it's working as expected.

After you have done repo init, make sure it has actually used the expected version of repo:

 $ cd .repo/repo
 $ git log

And make sure it's showing the expected commit, i.e. this change.

Then make sure the manifest is correct.

 $ cd ../manifests
 $ git log

Make sure it's showing the expected commit, i.e. the commit where you add the clone-depth option.

If you've made a local clone of the manifest and added a commit on that, you might find it's actually taken the remote branch rather than the local one, so your commit adding the clone-depth has not been used.

Also I noticed there's no revision specified on the git in your example manifest.  What is the default revision?  If it's a tag or sha1 I suspect the depth option will not work very well.

Patch-set: 2