Update patch set 5
Patch Set 5: Code-Review-1

I don't really understand the use case for this.

What happens for users who are mirroring a manifest that has path attributes, but they want the mirror to be created in the path according to the Git's name? This change will break repo for them.

Patch-set: 5
Label: Code-Review=-1