Allow mirror to be created in directories specified by 'path' attribute
In some cases, especially when local manifest files exist, users may want
to force the mirrored repositories to be created in folders according to
their 'path' attribute in the manifest, rather than according to the name
of the repositories.

Add a new option `--force-path` on the init command to support this.

Change-Id: Ib0a639e9014811701fbc4eccc3b03330d9b9f4b2
Signed-off-by: Scott Fan <fancp2007@gmail.com>