Update patch set 1
Patch Set 1:

erg...I don't like it in the manifest because I wouldn't want manifest authors who aren't the end user to use this feature, but at the same time, I don't want to expose this as an option to end users who are not the manifest author because they should not have to be burdened by something like this.

Something I was unsure of on the previous thread was are you just trying to deal with some hypothetical conflict?  Or is this a real problem that you are encountering?  If it's just a hypothetical problem, I'm not sure there's a real need to complicate things to address it.  If it's a real problem, ideally repo should be handling it automatically.

Patch-set: 1