Update patch set 1
Patch Set 1:

Sorry, I wasn't really clear with my use of the word "wrapper".  I understand that you are just sending commands via ssh.  What I mean is that the way you specify the command should not be different.  If querying is done directly with ssh host gerrit query "querystr" then the repo syntax should be repo gerrit query "querystr".

Another issue is that the ssh commands do not work on every gerrit instance (for example, gerrit-review and android-review).  However, the REST api which is getting more and more support will work even when ssh is blocked off, so I feel like this feature should work off of the REST api as opposed to the ssh api.

Patch-set: 1