Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

So the idea is that you init -b stable but then you push to master?  I'm a reluctant to continue shoving more and more attributes into Project and this use case seems like an arbitrary work flow.  I'd feel more comfortable with doing something like adding a --dest flag to the upload command (which would be useful in more usecases that just this one particular one).

Marking as -1 for now, but let me know if I've misunderstood.

Patch-set: 1