Fix pylint warning "Redefining built-in 'input'"
A couple of files have the following construct:

 try:
   input = raw_input
 except NameError:
   pass

which causes a pylint warning that the built-in name `input` is
being redefined.

The `raw_input` built-in seems to be available in most versions of
Python, so I can see no reason for this to be necessary.

Remove the redefinition and change the code to just use `raw_input`
directly.

Change-Id: Ia403e525b88d77640a741ac50382146e7d635924