Fix pylint warning "Redefining built-in 'input'"
A couple of files have the following construct:

 try:
   input = raw_input
 except NameError:
   pass

This was added because raw_input is not available in python3,
but it causes a pylint warning that the built-in name `input` is
being redefined.

Rework it so that `input` is not redefined.  Also add a comment
to clarify that the construct is needed to support python3.

Change-Id: Ia403e525b88d77640a741ac50382146e7d635924