Update patch set 3
Patch Set 3: Code-Review+1

Personally, I'd prefer doing # pylint-disable for these couple of lines.

The point of writing python2/3 compatible code this way is so that you can be writing legitimate python3 with the exception of a few lines at the top of the file that will be removed in the case support is ever removed for python2.  It's not that important, so I'll mark it as +1 and you can submit if you feel otherwise.

Patch-set: 3
Label: Code-Review=+1