Update patch set 2
Patch Set 2:

Fixed. auto_topic=True was left over from a merge I (incorrectly) performed with a branch on which we'd set the auto_topic option.

If you're interested, we're using the topic branches to shoehorn in unified cross-project code reviews. Each topic represents a set of changes that should be reviewed as a unit, and are tested as a unit by the CI system.

It would be really nice if Gerrit could allow reviewing topics as a cohesive whole instead of on a changeset-by-changeset basis...

Patch-set: 2
Reviewer: Gerrit User 1012614 <1012614@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0