New 'supplemental-metadata' manifest node
There is an issue with mixed-use of annotations and remove-project/project
pairs: removing a project removes its annotations such that they will need
to be duplicated when the project is re-added back into the manifest. This
duplication of information will quickly get out of control and result
in an inconsistent data model where a project's annotations in one place
are out of sync with its annotations in another place.

To work around this, we do not allow mixing of annotations with
remove-project/project pairs. Instead, we store all annotations for
all projects on the server in a special 'All-Projects' manifest. This
results in one and only one place for annotations to be specified. Good.

However, times may arise where we wish to do a repo forall over a tree
that was generated by a manifest that included remove-project/project
pairs and have at our fingertips all annotations for the given
projects. This is where this feature comes in: a manifest can now specify
one or more <supplemental-metadata> manifests which will pull in extra
metadata (tags & annotations) for each project.

Change-Id: I20b94e28cb044266c74fd4dbc3fe02bef70f5df6
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>