Update patch set 1
Patch Set 1: Code-Review-1

Yesssss....I've wanted to clean this up for a while.

However, I think it can be simplified even further.  When it was originally written the author thought it would be a good idea to support machines that couldn't support multiple threads, but that doesn't seem like a realistic use case we need to support.

I think it'd be best if we used a multiprocessing pool.  It'd get rid of the lock logic, the conditioned single thread scenario, and I think it would help with stuck ctrl-C interrupts (thread.daemon = True doesn't always seem to cut it).

So if you can update this to use a multiprocessing pool, that'd be awesome.

Patch-set: 1
Label: Code-Review=-1