repo: Support multiple branches for the same project.
It is often useful to be able to include the same project more than
once, but with different branches and placed in different paths in the
workspace. Add this feature. Fixes Issue 141.

This CL adds the concept of an object directory. The object directory
stores objects that can be shared amongst several working trees. For
newly synced repositories, we set up the git repo now to share its
objects with an object repo.

Each worktree for a given repo shares objects, but has an independent
set of references and branches. This ensures that repo only has to
update the objects once; however the references for each worktree are
updated separately. Storing the references separately is needed to
ensure that commits to a branch on one worktree will not change the
HEAD commits of the others.

One nice side effect of sharing objects between different worktrees is
that you can easily cherry-pick changes between the two worktrees
without needing to fetch them.

BUG=chromium:206946
TEST=repo {start,status,sync,upload} works with duplicate repositories
TEST=repo rebase -i project complains about duplicate repositories
TEST=Update an existing repo incrementally; check out a new repo.
TEST=TODO: Update cbuildbot to use the patched version of repo and test it
     on mario-paladin, x86-generic-full, and x86-mario-release.
CQ-DEPEND=CL:172893, CL:172913

Change-Id: Ib84e90bee8a2243b29b57257b833fade612bed74