Automatically setup alternates if syncing from a local tree
Detect that the repository being fetched from is on the local system,
and set up the alternates to point to the local repository.

My use case for this is when I've got a git mirror on my local machine
(not handled through repo --mirror), and my .gitconfig rewrites the
normal remote paths to the local mirror.

Change-Id: Icb86cac3d465552f30dd1c396561bf78dae6af78