Update patch set 2
Patch Set 2: Code-Review-1

Addressed your comments, David.

I think this is a situation where having unit tests would have helped.

Also, do we have any plan on the pickle version change?
A warning for that would be better than an Exception (could try it pickling a file first if one exists, and see if it's the correct version)

Patch-set: 2