Update patch set 1
Patch Set 1:

Initially it was a just quick hack to include it with repo. This is obviously becoming more difficult to manage.

I tried to avoid too much knowledge of gerrit in repo, but maybe we should consider migrating repo to fetching this hook from the configured review server instead of relying on the copy it bundles. But automatically installing the hook from gerrit should require the transport to be encrypted, e.g. SSH or https://, so there is reduced risk the user is talking to a potentially unknown host by mistake and installing something they didn't get a chance to review.

Patch-set: 1
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0