Update patch set 9
Patch Set 9:

Even better if it would update the hooks after a repo sync.
Since the hooks repo is part of the manifest as well, those hooks get updated after a sync, installing the updates after a sync would make sure the hooks being used are in sync with the hooks in the repo.

Patch-set: 9