Implementation of manifest defined githooks
This is based on change https://gerrit-review.googlesource.com/42236

When working within a team or corporation it is often
useful/required to use predefined git templates. This
change teaches repo to use a per-remote git hook template
structure.

The implementation is done as a continuation of the
existing projecthook functionality. The terminology is
therefore defined as projecthooks.

The downloaded projecthooks are stored in the .repo
directory as a metaproject separating them from the users
project forest.

The projecthooks are downloaded and set up when doing a
repo init and updated for each new repo init or a sync.

When downloading a mirror the projecthooks gits are
not added to the bare forest since the intention is to
ensure that the latest are used (allows for company policy
enforcement).

The projecthooks are defined in the manifest file in the
remote element as a subnode, the name refers to the
project name on the server referred to in the remote.
<remote name="myremote ..>
   <projecthook name="myprojecthookgit" revision="myrevision"/>
</remote>

The hooks found in the projecthook revision supersede
the stock hooks found in repo. This removes the need for
updating the projecthook gits for repo stock hook changes.

The projecthooks are updated within the sync operation prior
to any updates of the other git repos.

To test this change you will need to create a repo manifest
with a hooks repo included. When the hooks repo specified in
the manifest is empty the default hooks contained within repo
are used. Once a hook is added to the hooks repository a simlink
is put in place for the git repos.

Change-Id: I304cc6a63edfc0e7ccfc4c850876ad546d0c3395