Allow clone.bundle to redirect to a file:/// URL
When fetching from a company internal server it can be convenient to
make use of clone.bundle from a known location on the filesystem, e.g.
served over NFS.

Without this patch, curl fails with

	curl: (1) Protocol file not supported or disabled in libcurl

Curl accepts "--proto-redir +file" arguments even when built with
CURL_DISABLE_FILE (in which case they are a noop) so on systems with
new enough curl this should be safe.

The --proto-redir option was introduced in curl 7.21.0 (2010-04-24),
which is long enough ago that machines stuck on Debian oldstable have
it but recent enough that Ubuntu 10.04 (lucid) doesn't have it.  We
should check if curl is new enough before using the option.

Change-Id: I0a2cc5d1c1601fe258ff364dfd7338021a00fa66