Add support for Python3
execfile is not in python3. Add an if statement to switch between exec and execfile.

iter.next() is not in python. change to next(iter). This is backwards compatible to Python2.6.

dict.values(), .items(), and .keys() produce dict_* objects rather than list objects in Python3.
Wrap these to convert them to list objects.

map() produces a map object in Python3. Wrap these to convert them to list objects.

unicode has become str in Python3. If using python3, define unicode as str.

Change-Id: I2d8a66954bd2ba1f8b02916697b74dec6093a726