Add support for Python3
execfile is not in python 3. Add an if statement to switch between exec and execfile.

iter.next() is not in python. change to next(iter). This is backwards compatible to Python 2.6.

dict.values(), .items(), and .keys() produce dict_* objects rather than list objects in Python 3.
Wrap these to convert them to list objects.

map() produces a map object in Python3. Wrap these to convert them to list objects.

unicode has become str in Python 3. If using python 3, define unicode as str.

PS2:
enforce Protocol 2 on Pickle. Protocol 3 is not supported in Python 2.x.
This allows the user to switch between Python 2 and 3 without issues.

PS3:
Only warn about using Python 3 when running the repo script directly
This fixes the issue where the user is warned twice

Bug: Issue 117
Change-Id: I2d8a66954bd2ba1f8b02916697b74dec6093a726