Update patch set 1
Patch Set 1: Code-Review+1

No need for a converter function

If jsonFile points to a pickle file, then during the read process json will raise ValueError, which is caught, the file is removed and the function returns None.

As the function returns None, the file is is regenerated based on the git config into a json format.

No data is lost, no errors are uncaught

Patch-set: 1
Label: Code-Review=+1