Handle repository switches correctly.
In I5e2f4e1a7abb56f9d3f310fa6fd0c17019330ecd , repo was updated to store
each project in a different directory on disk. This introduced an issue
when a checkout is switched from one project to another: repo would
forget to update the symlinks on disk to point at the new location. This
in turn causes problems for repo because repo assumes that these symlinks are
correctly set up.

In this CL, we update repo to update the symlinks when repositories are
switched. This then introduces a second issue: repo can't update
existing checkouts because the objects are now missing. To fix this,
I've updated repo to reference the old objects using alternates,
allowing the checkout to be safely updated. In this case, old references
are blown away after the update via non-fast-forward updates. Some
references may be left around afterwards.

BUG=none
TEST=Test switching around what repository a given directory points to.
TEST=Test repo sync with old checkout.

Change-Id: Id50ff19ecd8b28769f8a683a5fb7d1b20f33e3a3