Don't mess with ref spec for existing repos
Currently, repo overwrites any modifications to remote.<remote>.fetch so
that it always looks like:

	fetch = +refs/heads/*:refs/remotes/<remote>/*

This can be bothersome when a user only wants to fetch a few branches
from the server, which they can do by tweaking this ref spec. Fix this
by not touching the fetch config for existing repos.

Change-Id: I8417a18d6bf57fa9d46fce2cb99dd4aefb27e61c