Update patch set 1
Patch Set 1:

If we were just working directly with a manifest repository, it'd be easy enough to do this using repo's -m option to select a manifest, and then just have that manifest selectively include various sub-manifests.  However, the problem is that we're working with a local_manifests repository, layered on top of an upstream manifest.  And unlike the manifest, there's no way to select specific local manifests, analogous to repo init -m.  This change would provide one way to solve that problem.

As an alternative, would you accept a change that added an option to parse a specific manifest from local_manifests?  local_manifests already parses all manifests in its top-level directory, but not any from subdirectories; would you accept a patch adding an option to parse a specific local_manifests subdirectory, or alternatively to parse a specific xml file in a local_manifests subdirectory?  "-l subdir/some-manifest.xml" would solve the problem just as well.

Patch-set: 1