Only avoid fetching from remotes in current-branch-only mode
In 2fb6466f795eb30c1dfa598501f5b5d2981e6a5f an optimisation was
added to avoid fetching from remotes if the project is fixed to
a revision and the revision is already available locally.

This causes problems for users who expect all objects to be
fetched by default.

Change the logic to only avoid fetching when using the "current
branch only" mode.

Change-Id: I3b2794ddd8e0071b1787e166463cd8347ca9e24f