Enable repo download to work in "multiple branches for the same project"
Fix includes:
1. "repo download" usage and help description modified. Now user needs
to supply the project path instead of the project name in repo download
command. As a result the change/patchset gets downloaded to the specified
path in the local workspace.

2. For the scenario of "multiple branches for the same project" if the
user enters the project name in the repo download command, then an error
message is displayed to the user to enter the path instead of the project.

Command Usage: repo download {target change[/patchset]}...

Bug: Issue 173
Change-Id: I4d07d110cd4252641469ad103bed9bffe55241d8