Update patch set 7
Patch Set 7:

(6 comments)

can it also be added on a per remote basis? perhaps one remote is http, and one remote is git.


Conley, a lot of these attributes are common, or should be common, between project, default and remote.

For the future, would it be a good idea to implement a class and routines for parsing/handling common attributes?

Patch-set: 7
Label: Verified=0