Update patch set 3
Patch Set 3:

> I'm not sure why it would given that we are mostly going back to old behavior.

I'm not sure I understand. The bug is seen in latest released repo, but now squashed from master. If "going back to old behavior" means we won't have a bug that is now seen in the current release, why is this a reason _against_ issuing a new release? I'm not saying we have too, just wonder about the reasoning.

Patch-set: 3