Only use sys.executable to run main.py when available
According to Python documentation, the sys.executable is not always
available. It could be an empty string or None in some cases, but it
has been used to run main.py since commit 53263d8.

This patch fixes the problem that main.py cannot be run when
sys.executable is empty, e.g in some sandbox environments, the Python
interpreter sets it as empty.

Change-Id: Ie7bd1eafba092eeab9a29051c49332ccede5e7c6