use explicit Python executable to run main.py when sys.executable is available
According to Python document, the sys.executable is not always available.
It could be empty string or None in some cases. but it has been used to run
main.py since revision 53263d8.

This patch fixes a problem that cannot run main.py by empty sys.executable.
(e.g - In some sandbox environment, Python interpreter set it as empty)
If it is not available to use, the main.py will be executed properly like
previous version of repo.

Change-Id: Ie7bd1eafba092eeab9a29051c49332ccede5e7c6