Update patch set 1
Patch Set 1: Code-Review-1

This was the behavior at some point, but some complained that they were expecting the whole branch to be there with -c, as had originally been the behavior.

The solution here is to set the upstream attribute of the project to a branch where the sha1 is found.

Patch-set: 1
Label: Code-Review=-1