Hold persistent proxy connection open while fetching clone.bundle
The persistent proxy may choose to present a per-process cookie file
that gets cleaned up after the process exits, to help with the fact
that libcurl cannot save cookies atomically when a cookie file is
shared across processes. We were letting this cleanup happen
immediately by closing stdin as soon as we read the configuration
option, resulting in a nonexistent cookie file by the time we use the
config option.

Work around this by moving the cookie logic into its own class with a
close method, so that we close the process only when we're done with
the cookie file.

Change-Id: I12a88b25cc19621ef8161337144c1b264264211a