Update patch set 1
Patch Set 1:

> This was the behavior at some point, but some complained that they
 > were expecting the whole branch to be there with -c, as had
 > originally been the behavior.
 > 
 > The solution here is to set the upstream attribute of the project
 > to a branch where the sha1 is found.

I see. How about fetching only the 

 > This was the behavior at some point, but some complained that they
 > were expecting the whole branch to be there with -c, as had
 > originally been the behavior.
 > 
 > The solution here is to set the upstream attribute of the project
 > to a branch where the sha1 is found.

I see. This will however not work if the sha1 is not the head of the upstream branch:
The first call to _RemoteFetch will fail and the following recursive call to _RemoteFetch will set current_branch_only to True (when depth is set) and there will be an infinite loop.

We could change the fallback behavior after the first failed RemoteFetch to:
* fetch the specific sha1 when depth is set
* if depth is not set, refetch with current_branch_only=True.

What do you think?

Patch-set: 1